requirements.txt
PyQt6>=6.6.0
mss>=9.0.0
soundcard>=0.4.5
numpy>=1.24.0
ffmpeg-python>=0.2.0
keyboard>=0.13.5
pillow>=10.0.0
qtawesome>=1.2.0
